requirements.txt
streamlit
huggingface-hub
python-dotenv
requests
pdfplumber
torch
transformers
numpy
nltk
